requirements.txt
PyMySQL==0.8.1
django_ckeditor==5.4.0
Django==2.0.5
